requirements.txt
flask-migrate
flask-sqlalchemy
flask-wtf
flask
pymysql
faiss-cpu
sentence-transformers
python-dotenv
